Upgrade metacat with enhancements to support external data linking
Enhances metacat build to patch the metacatui dataset XSL
to render external linking annotations in a separate
table. Addtionally, updates the json-ld view to transform
relation type annotations to Dataset.hasPart, Dataset.archivedAt,
Dataset.sameAs and Dataset.identifier

This modification pulls the latest eml-dataset.xsl from the
Metacat release patches it with eml-dataset.xsl.patch.

Closes ess-dive/ess-dive-project#196
Issue ess-dive/essdive-toolset#218

+++ metacat/docker-entrypoint.sh
@@ -94,12 +94,16 @@ if [ "$1" = 'bin/catalina.sh' ]; then
                 echo '**********************************************************'
                 echo "Copying base metacat skin properties files into ${skin_path}"
                 echo '***********************************************************'
-                echo
 
                 cp -Rv ${skin_path} ${METACAT_DIR}/style/skins/
 
-                cat $METACATUI_BASE_SKIN_PATH/metacatui.properties >> ${METACAT_DIR}/style/skins/${skin_name}/${skin_name}.properties
-                cp -v $METACATUI_BASE_SKIN_PATH/metacatui.properties.metadata.xml ${METACAT_DIR}/style/skins/${skin_name}/${skin_name}.properties.metadata.xml
+                if [ ! -f ${METACAT_DIR}/style/skins/${skin_name}/${skin_name}.properties ]; then
+                  cat $METACATUI_BASE_SKIN_PATH/metacatui.properties >> ${METACAT_DIR}/style/skins/${skin_name}/${skin_name}.properties
+                fi
+
+                if [ ! -f ${METACAT_DIR}/style/skins/${skin_name}/${skin_name}.properties.metadata.xml ]; then
+                  cp -v $METACATUI_BASE_SKIN_PATH/metacatui.properties.metadata.xml ${METACAT_DIR}/style/skins/${skin_name}/${skin_name}.properties.metadata.xml
+                fi
 
                 echo
                 echo '**********************************************************'
